// src/types/store.ts
export interface CommentStoreType {
  postId?: string;
  postAuthorId?: string;
  postTitle?: string;
  // screenName?: string;
  commentViewAddedToId?: string;
  commentViewAuthorId?: string;
  editComment?: CommentData;
  replyToComment?: CommentData;
  notiCommentId?: string;
  // origin?: string;
  content?: string;
  // receiverId?: string;
  receiverName?: string;
  // prefix?: string;
  // reset?: boolean;
}
// ...
export interface CommentData {
  id: string;
  addedToId: string;
  originalId: string;
  createdAt: string;
  commentContent: string;
  commentType: CommentType;
  commentDepth: CommentDepth;
  commentStatus: CommentStatus;
  commentAuthorId: string;
  commentAuthorName: string;
  commentReceiverId: string;
  numOfReported: number;
  numOfLikes: number | null;
  numOfSub: number;
  blockedBy: string[] | null;
  likedUsers: string[] | null;
  author: UserData;
  receiver: UserData;
  latestSubComment: CommentData | null;
  replyInfo: {
    repliedToId: string;
    repliedPostTitle: string;
  };
}

// export interface UpdateCommentInfo {
//   commentId: string | null;
//   contentToEdit: string | null;
//   receiverName?: string | null;
//   fromPostView?: boolean;
// }

export interface StoreAction {
  type: string;
  [key: string]: any;
}